Seperate doubly circular intrusive linked list into dclist.rs and change type parameter in spci.payload.
